+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Added sign message helper](https://github.com/multiversx/mx-sdk-dapp-core/pull/113)
 - [Fixed minor TODOs](https://github.com/multiversx/mx-sdk-dapp-core/pull/110)
 - [Fixed getWindowParentOrigin.test](https://github.com/multiversx/mx-sdk-dapp-core/pull/111)
 - [Updated managers to use internal folder](https://github.com/multiversx/mx-sdk-dapp-core/pull/108)